add redirect from host and add css and change host pages structure and fix vercel
