+++ src/app/pages/hosts/[hostId]/addHome/page.js
@@ -2,6 +2,7 @@ import { db } from "@/db";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import { auth } from "@clerk/nextjs";
+import "./addHome.css"
 
 import SaveHomeButton from "@/components/SaveHomeButton";
 export default async function AddHome({ params }) {
@@ -115,7 +116,7 @@ export default async function AddHome({ params }) {
           </div>
         </div>
 
-        <SaveHomeButton />
+        <SaveHomeButton className="buttOne" />
       </form>
     </div>
   );
