+++ src/app/pages/hosts/hosts.css
@@ -1,0 +1,34 @@
+.profileClass{
+    aspect-ratio: 1 / 1;
+    max-height: 20vh;
+    max-height: 20vh;
+}
+.bioCon{
+    display:flex;
+    flex-flow: row nowrap;
+    align-items: center;
+    justify-content: space-evenly;
+}
+.bio{
+    text-align: center;
+    font-size: large;
+    font-weight: bold;
+    font-family: 'Times New Roman', Times, serif;
+}
+.nameTitle{
+    margin: 0;
+    font-size: x-large;
+}
+.linkText{
+    letter-spacing:2px;
+    text-shadow:
+    0.07em 0 black,
+    0 0.07em white,
+    -0.07em 0 black,
+    0 -0.07em black;
+    background-color: rgba(00, 00, 00,0);
+    color: white;
+}
+.linkText:hover{
+    box-shadow: 0px 0px 3px 3px;
+}
